api/requirements.txt
# Core API
fastapi==0.115.2
uvicorn[standard]==0.29.0
pydantic==2.11.9
pydantic-settings==2.10.1

# Database
sqlalchemy==2.0.43
asyncpg==0.29.0
supabase==2.20.0
psycopg2-binary==2.9.10

# AWS
boto3==1.40.37
botocore==1.40.37

# Utilities
aiofiles==24.1.0
redis==6.4.0
python-multipart==0.0.20
python-dotenv==1.1.1
httpx==0.28.1